+++ src/spiceypy/tests/test_spiceerrors.py
@@ -27,7 +27,7 @@ import spiceypy as spice
 from spiceypy.utils.libspicehelper import libspice
 from spiceypy.cyice import cyice
 import spiceypy.found_catcher
-from spiceypy.tests.gettestkernels import cwd
+from spiceypy.tests.gettestkernels import cwd, CoreKernels, ExtraKernels
 import os
 
 
@@ -92,16 +92,32 @@ def test_no_loaded_files_exception():
     with pytest.raises(spice.SpiceyError):
         spice.ckgp(0, 0, 0, "blah")
     spice.reset()
+    spice.kclear()
+    with pytest.raises(spice.exceptions.SpiceNOLOADEDFILES):
+        spice.ckgp(0, 0, 0, "blah")
+    spice.reset()
+    spice.kclear()
+    spice.furnsh(CoreKernels.testMetaKernel)
+    spice.furnsh(ExtraKernels.v1jCk)
     with pytest.raises(spice.NotFoundError):
         spice.ckgp(0, 0, 0, "blah")
     spice.reset()
-    # with pytest.raises(spice.SpiceyError):
-    #     cyice.ckgp(0, 0.0, 0, "blah")
-    # spice.reset()
-    # with pytest.raises(spice.NotFoundError):
-    #     cyice.ckgp(0, 0.0, 0, "blah")
-    # spice.reset()
-    with spiceypy.found_catcher.no_found_check():
+    spice.kclear()
+    with pytest.raises(spice.SpiceyError):
+        cyice.ckgp(0, 0.0, 0, "blah")
+    spice.reset()
+    spice.kclear()
+    with pytest.raises(spice.exceptions.SpiceNOLOADEDFILES):
+        cyice.ckgp(0, 0.0, 0, "blah")
+    spice.reset()
+    spice.kclear()
+    spice.furnsh(CoreKernels.testMetaKernel)
+    spice.furnsh(ExtraKernels.v1jCk)
+    with pytest.raises(spice.NotFoundError):
+        cyice.ckgp(0, 0.0, 0, "blah")
+    spice.reset()
+    spice.kclear()
+    with spiceypy.no_found_check():
         with pytest.raises(spice.SpiceyPyIOError):
             spice.ckgp(0, 0, 0, "blah")
         spice.reset()
@@ -137,7 +153,7 @@ def test_error_to_str():
 
 def test_disable_found_catch():
     spice.kclear()
-    with spiceypy.found_catcher.no_found_check():
+    with spiceypy.no_found_check():
         name, found = spice.bodc2n(-9991)
         assert not found
     with pytest.raises(spice.SpiceyError):
@@ -158,7 +174,7 @@ def test_recursive_disable_found_catch():
         if i <= 0:
             return
         else:
-            with spiceypy.found_catcher.no_found_check():
+            with spiceypy.no_found_check():
                 name, found = spice.bodc2n(-9991)
                 assert not found
                 _recursive_call(i - 1)